Don't endup in a endless loop for inline signed messages
For inline signed messages the reference node is a null pointer
so we also need to set the memento for this null sign pointer, to
be able to refind the memento again.

This patch makes it impossible to have multiple inline signed messages blocks.
But one the other side it wasn't working before, too :)

BUG: 356093
REVIEW: 126451