Remove moveCompletionPopup() hack.
Summary:
1) commit b78a77e95659 in kdelibs 4.9.2 (see bug 270473) takes
care of making KCompletionBox move/resize when the parent widget has moved/resized.

2) this hack made the completion popup not appear for me, the very first time.
After hide()+show(), isVisible() was true, but the popup wasn't actually visible
(geometry was correct though). Possibly a Qt issue, but that's what we get when
doing hacks.
Worse, this broke the Tab key to focus the next field, since the code thought
it was about tabbing through the presumed-to-be-visible completion box.