Fix giving focus back to the messagelist from the search lineedit.
Typing a filter and then clicking back into the messagelist (to be able to use
Left/Right or Del shortcuts for instance) only worked when clicking on a
*different* current item (because slotSelectionChanged ends up calling setFocus).
Clicking on the existing current item had no effect.

The simplest fix is to just use ClickFocus as focus policy, but
of course tabbing should work too, so rather StrongFocus, which is
actually the default, so just removing the line is best.

In kdepim4 the default focus policy was used too, c626f4c added this
NoFocus line by error: it doesn't actually fix the dual-color-issue for
selected messagelist items.