Allow only one recipient per line in RecipientsEditor
For the opportunistic encryption we want to show an icon next to
each recipient showing status of the encryption key we are going
to use. It is however not possible to put multiple icons mid-text
into K(Q)LineEdit, only at the beginning of the line which is a
problem if there are multiple recipients in a single line. The
RecipientsEditor already jumps to next line automatically when you
hit select a contact from autocompletion, so this patch applies
the same behaviour to manual input as well. Just typing a comma
(or semicolon) or pasting a comma-separeted list of recipients
into the recipient edit will automatically split the input to
multiple lines, having one recipient per line.

Differential Review: https://phabricator.kde.org/D2313