Make message switching in kmail much faster (also fixes a focus issue)
This call to load() triggers a lot of internal stuff in WebEngine,
including NavigateToEntry, creation (and soon afterwards, deletion) of
QtWebEngineCore::RenderWidgetHostViewQtDelegateWidget, etc.

I can't see any regression when removing this call.

On the contrary, this fixes a focus issue: when clicking on an
external link in an email and coming back to kmail, arrow up/down
didn't work anymore, because the last widget installed as focusproxy
for the webview (the DelegateWidget above) had been deleted meanwhile.

Differential Revision: https://phabricator.kde.org/D2637