MessageList: load icons in the size that is actually used by the Theme
Improves looks and performance. Previously we had a single cache of icon
pixmaps in Manager. Manager always loaded 64x64 icons, even if the theme
was using 16x16 (which is the default). That resulted in ugly icons, both
due to down-scaling while rendering, and because we were using icons
designed to be shown as 64x64, not 16x16 or similar. This change moves
the pixmap cache to Theme, so that each Theme has its own cache using
the Theme's icon sizes. The cache is populated lazily, so unused
themes don't waste memory with cache.

It also slightly improves performance, because the pixmaps don't have
to be scaled down when rendering.