Optimize Item::indexOfChildItem() cache miss
If mThisItemIndexGuess is out of date then we will search for the item
in both directions from the old guess. Since in most cases the item has
only moved a few rows up or down, searching from the old guess means we
find the item faster than if we would just search from the beginning of
the container.