Bring back some primitive processing of HTML email content
We lost this entirely as part of the WebKit -> WebEngine port, emitting
the entire HTML content including <html> elements, headers, etc. into the
final document. WebEngine is fortunately tolerant enough to not choke on
that, but our tests do. So add a primitive and incomplete way to extract
the body, until we get a proper solution back. That's still better than
nothing, and it unbreaks the render test.