Centrally add global Grantlee context properties
Summary:
This way the individual render methods or plugins wont have to deal with
this anymore, and the Grantlee template become easier to re-use.

The next step is to replace the ad-hoc block properties with directly
exported MessagePart instances where possible, they are QObjects and even
have some of the needed properties already.

Reviewers: knauss

Reviewed By: knauss

Subscribers: #kde_pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D8123