Patch from aleksejshilin
When no default aggregation or theme is selected by user, Core::Manager

picks the first one. The available ones were stored in QHash containers
which, per Qt documentation, are ordered randomly:

When iterating over a QMap, the items are always sorted by key. With
QHash, the items are arbitrarily ordered.

Given that keys are generated in a way that they differ each time, the
result of picking the first item was unpredictable.

This commit changes containers' type to QMap which guarantees order.

Differential Revision: https://phabricator.kde.org/D10417