KMail: fix selecting a favorite that is filtered out in the foldertreeview.
Summary:
Instead of reacting to a selection change in the foldertreeview's
selection model (which won't happen, if the newly selected folder is
filtered out in that view), let's make Pane react on the explicit call
to setCurrentFolder. For this we need the QModelIndex in the arguments
to that method (=> corresponding kmail change coming up).

Side fix: set mPreferEmptyTab to false in case setCurrentFolder is
called twice; but in the end I fixed that in
KMMainWidget::slotFolderChanged so that it's not called twice.

Test Plan:
Type "zzz" in the filter-lineedit for the folder-tree-view,
so everything disappears. Then try to select a favorite folder.

Reviewers: vkrause, dvratil, mlaurent

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D14139