Thread messages with same reference ID together
Summary:
Thread messages with same reference ID together

Messages from github or bugzilla usually use In-Reply-To and References headers
with an ID that either doesn't actually exist or just isn't around in the
folder. In case of bugzilla, no threading happens in that case because subject
lines don't start with "Re:". github does do that, but oftentimes the subject
based threading is imprecise (by design).

Therefore create a sorted-by-date cache of messages similar to the subject
based cache and put any message we cannot perfectly thread to the very first
message with the same message ID.
What results is a flat but threaded messagelist similar to what thunderbird does.

FIXED-IN: 5.10.0
BUG: 279115

Reviewers: dvratil

Reviewed By: dvratil

Subscribers: dvratil, bruns, kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D14883