Move management of QWebEngineProfile to WebEnginePage
To avoid a crash on KMail quit or viewer window close, caused by the
MailWebEnginePage being deleted while its QWebEngineProfile is still
active.  Simplified management of the profile by handling its creation
and deletion within WebEnginePage.

Deprecate the 2-argument WebEnginePage constructor which allows an
already allocated profile to be used.  This constructor is used by
Akregator which specifies the global profile; it will be changed to
use a private profile instead.

Differential Revision: https://phabricator.kde.org/D19559