Generalize the HTML warning box for inline messages
Summary:
This follows what KMessageWidget or Kirigami.InlineMessage do, making this
visually and conceptually more consistent with everything else.

In here we only need the error message box (for the HTML warning), the
other ones are going to be used by plugins.

While at it, align the visual style for <hr> with Kirigami.Separator.

Reviewers: knauss, mlaurent

Reviewed By: knauss, mlaurent

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D23299