tests(mimetreeparser): Add tests for the MessagePartTree.
Summary:
The MessagePartTree out of a mail is the interface of MimetreeParser,
so we should have tests, that garantee that the tree is correctly
created.
Currently this test is also done on rendertest (MessageViewer), but
there many more things are tested, we can't remove the code there for
the moment.

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D23397