feat(mimetreeparser): to get rid of rendering in parsing library completly.
Summary:
MimeTreeParser is about to parse mails and not render them, remove the
rendering part completly from the API to move the rendering
to MessageViewer.

That patch finally makes it possible to display MemoryHole headers in
MessageViewer.

This patch needs also a counterpart on kdepim-addons, as we also need to
update tests there.

Test Plan: run autotests.

Reviewers: dvratil

Subscribers: kde-pim

Tags: #kde_pim

Maniphest Tasks: T742

Differential Revision: https://phabricator.kde.org/D23807