Make *DefaultMessageColor use semantically correct colorscheme colors
Summary:
The unread message color was changed to ActiveText and the important message color was changed to NeutralText.

After looking at the documentation in the KColorScheme header file, it seemed to me that ActiveText and NeutralText were more appropriate than Complementary FocusColor and NegativeText. ActiveText is appropriate for new items and Neutral Text is appropriate for warnings. Obviously, I can't use ActiveText for both and NegativeText is for errors, which important messages are not, so I used NeutralText for important messages.

This currently results in almost no visual changes except for unread messages being a slightly brighter shade of blue and important messages being orange. In the future, I will change ActiveText in the Breeze colorscheme to yellow.

Test Plan:
old
{F7432990}
new
{F7432980}

Reviewers: #vdg, #kde_pim_kmail, ngraham, mlaurent

Reviewed By: #vdg, ngraham, mlaurent

Subscribers: ngraham, mlaurent, kde-pim

Tags: #kde_pim, #vdg

Differential Revision: https://phabricator.kde.org/D24230