Fix reported bug about replies to self
Summary:
Normally, replies are sent to the sender of the original message.
However, if the KMail user is the sender of the original message, KMail
sends the reply to the recipient of the original message, on the assumption
that the reply is meant to add information to the original.

Bugs 301449, 366356, and 392433 point out that this is not ideal if the
sender and recipient are two different identities of the KMail user.  If
the user sends mail from "work.id@example.com" to "home.id@example.org",
the reply should go from "home.id@example.org" to "work.id@example.com".

Bug: 366356
Bug: 392433
Bug: 301449

Reviewers: #kde_pim, mlaurent, knauss

Reviewed By: #kde_pim, knauss

Subscribers: mlaurent, knauss, kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D24035