Ignore Reply-To when replying to your own messages
Summary:
If you reply to an email that you sent to someone else, the ReplySmart
strategy sends a reply from yourself to the previous recipients, since
you probably intended to create a follow-up to the original.

However, ReplySmart should ignore any Reply-To header in the original;
otherwise it sends the reply back to yourself.

The first of the new test cases tests this situation.  The second new
test shows that the code changes did not break Reply-To handling for
original messages that came from a mailing list.

Reviewers: knauss

Reviewed By: knauss

Subscribers: knauss, kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D24302