fix(mesasgecomposer): encrypted and signed message like encrypted
Summary:
From the message structure a signed&encrypted message should look like a
simple encrypted message, to not leak any information to the outside
world. This also make the tests for SignEncryptJob passes.

No the output for OpenPGPMime is a mail, that has the signed only content
in first mime part and the encrypted mail in the second part. This is for
sure not that what we want.

Test Plan: run autotests successfuly.

Reviewers: #kde_pim

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D28728