Fix assert when deleting a folder in kmail
MessageList::Pane::Private::setCurrentFolder ends up calling
StorageModel::savePreSelectedMessage for the old folder, usually in
order to save the current message in that folder. But if we just deleted
that folder (and that's the reason we're "moving away" from it), then
it's not selected anymore and StorageModel::id() returns empty.

Just return early from savePreSelectedMessage in that case.