Fix[messagecomposer]: Do copy all mail headers instad of reference them.
When dealing with encryption and protected headers it is enough to
reference the headers in the messagepart, as the result is encrypted
directly. This is different for Sign only and there may be jobs after
signing that are changing the mail headers, so we need to copy all
headers into the encapsulated part instead of referencing them.

BUG: 427091
FIXED-IN: 5.16.1