Revert "Fix more mem leak found by ASAN"
This reverts commit 4fff608d0bde927ac2aac61d3afedbfec4bb4b80.

These deletion of those headers make EncrytJob, SignEncryptJob and
SignJob fail, because they are async and the ~SinglePartJob is called
when gpg is doing it work. And it is already visible in the test suite
MultipartJobTest and SignJobTest both failing because of this.
Furthermore we should not delete only some headers, instead we should
delete those headers when we delete the complete message.

CCMAIL: montel@kde.org