fix tests that break when run with QTEST_GUILESS_MAIN.
MessageComopser::UtilTests and MessageCore::StringUtilTests depend on
MessageCoreSettings, that depends on ColorUtils that needs access to the
palette and this is in GUI depdendend. And this means it needs
QTEST_MAIN to run successfully.

CCMAIL: montel@kde.org

This reverts part of commit 5e2aafedc308d51b8e574354d5d60b88bd9dc61d.