[RFC]: Split AttachmentControllerBase into a core class and a widget class
All the QObject only stuff has been moved to
AbstractAttachmentControllerBase and AttachmentControllerBase inherit
from it. This allow to use the class from a QML application while
sharing the logic.

KMessageBox::sorry has been replaced by signal emmited by the abstract
class and the base class listen to it to show the old dialog.

This is a RFC because the end goal is to make the ComposerViewer class
not depend on QWidgets anymore, so it will need some similar changes to
other classes. I want to make sure this change will be accepted after
being cleaned up.