Allow deletion of attachments from message structure view
Deletion of attachments (and other message parts) is only allowed
one at a time. Message parts inside encrypted message parts cannot
be deleted.

This (re-)adds the "Delete Attachment" action to the context menu
of the message structure view. It was removed 3 years ago because
it always crashed (because it tried to work with a copy of the node
tree of the original message).

Sponsored by: Intevation GmbH

CCBUG: 287177
BUG: 436089
FIXED-IN: 5.18.0