Fix deletion of two (or more) attachments
Deleting the first attachment triggers an update of the message. In case
of IMAP this is done by adding a new message and deleting the old one.
This changes the remote id of the message. The problem was that after
deleting the second attachment the update of the message tried to set
the old remote id. This fails because only the server can set a remote id.
This problem is fixed by clearing the remote id before starting an
ItemModifyJob.

Funded by: Intevation GmbH

BUG: 282792
FIXED-IN: 5.18.0