Show the compliance status of signed messages (if needed)
If GnuPG enforces compliance, then show the compliance status of the
verification of the signature. We use "signWarn" for non-compliant
signatures.

The {% with space as " " %} hack is necessary to get whitespace between
the signature status and the compliance value. The template engine
removes any literal whitespace (which is often desired).

GnuPG-bug-id: 6343