Fix fallback path in MessageFactoryNG::applyCharset
In the case that the codec of the original message could not encode the reply,
it was still set as charset but the body encoded with the fallback codec.
This resulted in replies having messed up encoding.
It can be triggered by replying to multipart mails which define the charset
in parts only or if the reply template ends up with other special characters.

BUG: 447297
BUG: 443009
BUG: 298349