Look for matching subkey if no key was found for fingerprint
If the message was signed with a signing subkey instead of with the
primary key of an OpenPGP certificate, then we won't find a key with
findByFingerprint(). To look for a matching subkey we need to use
findSubkeysByKeyID().

FIXED-IN: 5.23.1
BUG: 469304
(cherry picked from commit 606ea1478d2d5b5aacdc6ef3f050655fe0352d87)