Fix extracting names with a comma from addresses in snippets
The old implementation splits the list of addresses at comma
characters. However, the display name of a mail address may contain a
(quoted) comma character as well in the form "Last, First".
Unfortunately, the old implementation does not respect quotations.

This change replaces the direct use of split by
KEmailAddress::splitAddressList, which is designed to do the right
thing.