Make the method param in iTIP mime part content-type configurable
The iMIP RFC says that the "method" param in the Content-Type
header of the mime part containing an iTIP invitation must be
the same as the value of the METHOD parameter in the iCalendar
itself.

This code is not aware of the details of iCal, so we provide the
method as a string alongside other parameters.