Port templateparser away from KMime::addContent
All cases here operate on newly created multipart nodes, so this is
straightforward to port. The warning in the docs about the dangerous
automatic multipart transformation of KMime::addContent are obsolete
with that method going away.

There is one call to KMime::removeContent left though that needs further
investigation as it's not yet clear what that operates on exactly.