Port NodeHelper away from deprecated Content::add/removeContent
The new methods do the same, just without the automatic multipart
transformation. It's somewhat tricky to prove here that this never
hits the case of a multipart transform unfortunately.

The extra nodes are only used for expanding signed/encrypted parts
(and thus always have a different multipart parent type), and there
is quite some unit test coverage on this as well.