Remove Util::allContents()
We only use this once to get the count of all KMime nodes in the tree,
but we never actually use the full flattened list of all nodes. So instead
just count directly, which is more efficient.

This also helps with upcoming const-correctness fixes in KMime which
conflict with Util::allContents (or rather the NodeHelper methods it uses)
dropping const-ness.