Avoid allocations in MD5 hash computation
This is on the hot path for building up the message list threading.

MD5 hashes are just 16 bytes, we can store that in a corresponding value
rather than heap allocating QByteArrays for them.

As a bonus this also makes the message list item object smaller on 64bit
systems as the "outer" part of QByteArray grew in size in Qt 6.