Optimize automatic expansion of message list view items
Don't animate the expansion - that is virtually imperceptible
while loading a folder anyway.

This helps with loading of folders containing thousands of unread
messages - I have some mailing list folders like that. Unread
conversations are automatically expanded, which I found to be
moderately expensive in time (a few %) and very expensive in
memory (>60% of KMail total).
The heaptrack trace is ModelPrivate::attachMessageToParent(),
[...], QTreeViewPrivate::expand(), then two calls of
QImageData::create() in two sub-calltrees 10 and 18 deep. 5533
allocations, 1.4 GB(!) peak memory usage. That part of peak memory
usage disappears with this patch.

My best guess is that it's a pathological case where a large
portion of the view is cached in a QPixmap while its displacement
is animating. What it can't be is QTreeViewPrivate::
AnimatedOperation instances (and so QPixmap instances) piling up:
a view can only have one AnimatedOperation at a time.
In any case, animations are somewhat expensive, and isAnimated()
and setAnimated() calls are very cheap, so this should be a net
benefit even if I saw mainly measurement artifacts.
Should be fun to investigate if anyone feels like doing that.