Fix "scrolling lock" to the bottom of the message list view
It didn't work for me in some folders. At the time that the
previous verticalScrollBar()->value() was taken, minimum and
maximum scroll positions were equal (both zero) and so value() was
zero, too, interpreted as top scroll position.
"Restoring" that position did not give the desired result.
Use the sort direction as a tie-breaker instead and *return* the
scrolling lock position instead of guessing (badly) in the caller.

Additonal observation: the notFirstLoadingJob condition is
strange: it was supposed to keep the list scrolled at the top or
bottom *only for e-emails arriving after the startup phase* - it
doesn't seem to have had that effect anymore and might instead have
sometimes prevented the bug that this commit tries to fix properly.
Might be possible to remove it now and / or to otherwise restore
the old intended behavior for people who want that.
Reference: commit c4fea23364fc175c97396a5d3506ebbfa2803de8 in
kdepim / svn revision 978098.

CCBUG: 264297