Fix "full-quick" search in message list for unread
("full-quick" is the quick search line with the detailed flags UI
opened, as opposed to just pressing the unread tool button in the
normal / small size quick search bar)

Unread is normally the absence of Read, which the matching
machinery is unable to match without special handling.
It could be done better with just slightly more advanced bit mask
logic (which would include a care / don't care mask) for filtering
on flags, comparing all flag bits in a few machine instructions,
but it's usually just one flag and it seems fast enough.